+++ src/main/java/video/api/java/sdk/domain/analytic/analyticEvent/SessionEventAnalyticsClient.java
@@ -3,14 +3,9 @@ package video.api.java.sdk.domain.analytic.analyticEvent;
 import video.api.java.sdk.domain.QueryParams;
 import video.api.java.sdk.domain.exception.ResponseException;
 
-import java.util.Iterator;
-
 
 public interface SessionEventAnalyticsClient {
+    Iterable<AnalyticEvent> list(String sessionId) throws ResponseException;
 
-    Iterator<AnalyticEvent> list(String sessionId) throws ResponseException, IllegalArgumentException;
-
-    Iterator<AnalyticEvent> search(QueryParams queryParams, String sessionId) throws ResponseException, IllegalArgumentException;
-
-
+    Iterable<AnalyticEvent> search(QueryParams queryParams, String sessionId) throws ResponseException;
 }