+++ src/main/java/video/api/java/sdk/infrastructure/pagination/IteratorIterable.java
@@ -1,0 +1,16 @@
+package video.api.java.sdk.infrastructure.pagination;
+
+import java.util.Iterator;
+
+public class IteratorIterable<T> implements Iterable<T> {
+    private Iterator<T> iterator;
+
+    public IteratorIterable(PageIterator<T> iterator) {
+        this.iterator = iterator;
+    }
+
+    @Override
+    public java.util.Iterator<T> iterator() {
+        return iterator;
+    }
+}