+++ CHANGELOG.md
@@ -9,4 +9,5 @@ All notable changes to this project will be documented in this file.
 - Move UploadProgressListener to domain
 - Use interfaces in Client
 - Replace PageIterator references with Iterator in domain
+- Replace Iterator with Iterable in domain interfaces